feat(react): Measure checkout request
