+++ react/src/components/Checkout.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
+import measureRequestDuration from '../utils/measureRequestDuration';
 import './checkout.css';
 import * as Sentry from '@sentry/react';
 import { connect } from 'react-redux';
@@ -40,6 +41,7 @@ function Checkout(props) {
 
   async function checkout(cart) {
     Sentry.metrics.increment('checkout.click');
+    const stopMeasurement = measureRequestDuration('/checkout');
     const response = await fetch(props.backend + '/checkout?v2=true', {
       method: 'POST',
       headers: { 'Content-Type': 'application/json' },
@@ -50,6 +52,9 @@ function Checkout(props) {
     }).catch((err) => {
       Sentry.metrics.increment('checkout.error', 1,  { tags: { status: 500 } });
       return { ok: false, status: 500 };
+    }).then((res) => {
+      stopMeasurement();
+      return res;
     });
     if (!response.ok) {
       Sentry.metrics.increment('checkout.error', 1,  { tags: { status: response.status } });
