+++ react/src/components/Products.js
@@ -3,6 +3,7 @@ import './products.css';
 import * as Sentry from '@sentry/react';
 import { connect } from 'react-redux';
 import { setProducts, addProduct } from '../actions';
+import measureRequestDuration from '../utils/measureRequestDuration';
 import Loader from 'react-loader-spinner';
 import ProductCard from './ProductCard';
 import { useState, useEffect } from 'react';
@@ -95,7 +96,7 @@ function Products({ frontendSlowdown, backend }) {
       // because it returns product data with a fast backend response.
       // Otherwise use the /products endpoint, which provides a slow backend response.
       const productsEndpoint = determineProductsEndpoint();
-      const start = Date.now();
+      const stopMeasurement = measureRequestDuration(productsEndpoint);
       fetch(backend + productsEndpoint, {
         method: 'GET',
         headers: { 'Content-Type': 'application/json' },
@@ -116,14 +117,10 @@ function Products({ frontendSlowdown, backend }) {
         .then(renderProducts)
         .catch((err) => {
           return { ok: false, status: 500 };
-        }).then(() => {
-          const end = Date.now();
-          const duration = end - start;
-          Sentry.metrics.distribution('request.duration', duration, {
-            unit: 'milliseconds',
-            tags: { endpoint: productsEndpoint }
-          });
-        })
+        }).then((res) => {
+          stopMeasurement()
+          return res
+        });
     }
 
     getProducts(frontendSlowdown);
